relocate shelter, pet, animal_tipe views and serializers. add subscription to shelters
