+++ help_paw/info/models.py
@@ -2,6 +2,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.db.models.signals import m2m_changed
 from django.dispatch import receiver
+
 from shelters.models import Shelter
 
 MAX_IMAGE_SIZE = 5 * 1024 * 1024
